pyproject.toml
[tool.poetry]
name = "hbim-open-api-client"
version = "0.1.0"
description = "A client library for accessing HBIM-OpenAPI"
authors = []
readme = "README.md"
packages = [
    {include = "hbim_open_api_client"},
]
include = ["CHANGELOG.md", "hbim_open_api_client/py.typed"]


[tool.poetry.dependencies]
python = "3.11.7"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
